requirements.txt
django-autoslug-iplweb
pillow
bootstrap4
